fixed single page

+++ admin/update_product.php
@@ -1,11 +1,8 @@
 <?php
 
-include '../components/connect.php';
-
+include '../connection.php';
 session_start();
-
 $admin_id = $_SESSION['admin_id'];
-
 if(!isset($admin_id)){
    header('location:admin_login.php');
 };
@@ -16,34 +13,44 @@ if(isset($_POST['update'])){
    $name = $_POST['name'];
    $price = $_POST['price'];
    $category = $_POST['category'];
+   // Update product information
+   $update_product_query = "UPDATE `products` SET name = '$name', categoryID = '$category', price = '$price' WHERE id = '$pid'";
+   $update_product_result = mysqli_query($conn, $update_product_query);
 
-   $update_product = $conn->prepare("UPDATE `products` SET name = ?, category = ?, price = ? WHERE id = ?");
-   $update_product->execute([$name, $category, $price, $pid]);
-
-   $message[] = 'product updated!';
+   if($update_product_result) {
+      $message[] = 'Product updated!';
+   } else {
+      $message[] = 'Error updating product: ' . mysqli_error($conn);
+   }
 
    $old_image = $_POST['old_image'];
    $image = $_FILES['image']['name'];
-   $image = filter_var($image, FILTER_SANITIZE_STRING);
+   // $image = filter_var($image, FILTER_SANITIZE_STRING);
    $image_size = $_FILES['image']['size'];
    $image_tmp_name = $_FILES['image']['tmp_name'];
    $image_folder = '../uploaded_img/'.$image;
 
    if(!empty($image)){
       if($image_size > 2000000){
-         $message[] = 'images size is too large!';
+         $message[] = 'Images size is too large!';
       }else{
-         $update_image = $conn->prepare("UPDATE `products` SET image = ? WHERE id = ?");
-         $update_image->execute([$image, $pid]);
-         move_uploaded_file($image_tmp_name, $image_folder);
-         unlink('../uploaded_img/'.$old_image);
-         $message[] = 'image updated!';
+         // Update product image
+         $update_image_query = "UPDATE `products` SET image = '$image' WHERE id = '$pid'";
+         $update_image_result = mysqli_query($conn, $update_image_query);
+         if($update_image_result) {
+            move_uploaded_file($image_tmp_name, $image_folder);
+            unlink('../uploaded_img/'.$old_image);
+            $message[] = 'Image updated!';
+         } else {
+            $message[] = 'Error updating image: ' . mysqli_error($conn);
+         }
       }
    }
 
 }
 
 ?>
+
 
 <!DOCTYPE html>
 <html lang="en">
@@ -60,7 +67,7 @@ if(isset($_POST['update'])){
 </head>
 <body>
 
-<?php include '../components/admin_header.php' ?>
+<?php include './admin_header.php' ?>
 
 <!-- update product section starts  -->
 
@@ -69,12 +76,12 @@ if(isset($_POST['update'])){
    <h1 class="heading">update product</h1>
 
    <?php
-      $update_id = $_GET['update'];
-      $show_products = $conn->prepare("SELECT * FROM `products` WHERE id = ?");
-      $show_products->execute([$update_id]);
-      if($show_products->rowCount() > 0){
-         while($fetch_products = $show_products->fetch(PDO::FETCH_ASSOC)){  
-   ?>
+$update_id = $_GET['update'];
+$show_products_query = "SELECT * FROM `products` WHERE id = '$update_id'";
+$show_products_result = mysqli_query($conn, $show_products_query);
+if(mysqli_num_rows($show_products_result) > 0){
+   while($fetch_products = mysqli_fetch_assoc($show_products_result)){  
+?>
    <form action="" method="POST" enctype="multipart/form-data">
       <input type="hidden" name="pid" value="<?= $fetch_products['id']; ?>">
       <input type="hidden" name="old_image" value="<?= $fetch_products['image']; ?>">
@@ -84,13 +91,23 @@ if(isset($_POST['update'])){
       <span>update price</span>
       <input type="number" min="0" max="9999999999" required placeholder="enter product price" name="price" onkeypress="if(this.value.length == 10) return false;" class="box" value="<?= $fetch_products['price']; ?>">
       <span>update category</span>
-      <select name="category" class="box" required>
-         <option selected value="<?= $fetch_products['category']; ?>"><?= $fetch_products['category']; ?></option>
-         <option value="main dish">main dish</option>
-         <option value="fast food">fast food</option>
-         <option value="drinks">drinks</option>
-         <option value="desserts">desserts</option>
-      </select>
+            <select name="category" class="box" required id="box">
+                <?php
+                $show_category_query = "SELECT * FROM `categories`";
+                $result = mysqli_query($conn, $show_category_query);
+
+                // Check if any categories were found
+                if ($result && mysqli_num_rows($result) > 0) {
+                    while ($row = mysqli_fetch_assoc($result)) {
+                        $name = $row['name'];
+                        $id=$row['id'];
+                        echo "<option value=\"$id\">$name</option>";
+                    }
+                } else {
+                    echo '<option value="" disabled selected>No categories found</option>';
+                }
+                ?>
+            </select>
       <span>update image</span>
       <input type="file" name="image" class="box" accept="image/jpg, image/jpeg, image/png, image/webp">
       <div class="flex-btn">
